Split decision_variable and constraint into separate files
