+++ python/ommx/ommx/_ommx_rust.pyi
@@ -158,6 +158,9 @@ class DecisionVariable:
         upper: typing.Optional[builtins.float] = None,
         name: typing.Optional[builtins.str] = None,
     ) -> DecisionVariable: ...
+    @staticmethod
+    def decode(bytes: bytes) -> DecisionVariable: ...
+    def encode(self) -> bytes: ...
     def __repr__(self) -> builtins.str: ...
 
 class DecisionVariableAnalysis:
